osa4 blogilista 4.21

+++ Osa4/Blogilista/controllers/blogs.js
@@ -27,6 +27,7 @@ blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({
       error: 'token missing or invalid'
@@ -50,8 +51,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const userId = decodedToken.id
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+  response.status(401).json({
+    error: 'invalid token'
+  })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
